Add Tooltip messages as mixin for now
