+++ app/javascript/mentor/DashboardMenu.vue
@@ -59,6 +59,7 @@
 <script>
 import { createNamespacedHelpers } from 'vuex'
 import menuMixin from 'mixins/menu'
+import tooltipsMixin from 'mixins/tooltips'
 
 const { mapGetters } = createNamespacedHelpers('authenticated')
 
@@ -68,18 +69,12 @@ import RegistrationMenu from 'registration/DashboardMenu'
 import JudgingMenu from './menus/Judging'
 import TeamMenu from './menus/Team'
 
-const Tooltips = {
-  MUST_SIGN_CONSENT_ON_TEAM: 'You must be on a team and sign a consent waiver to work on submissions',
-  MUST_SIGN_CONSENT: 'You must sign a consent waiver to work on submissions',
-  MUST_BE_ON_TEAM: 'You must be on a team to work on submissions',
-  AVAILABLE_LATER: 'This feature will open later in the Season',
-}
-
 export default {
   name: 'dashboard-menu',
 
   mixins: [
     menuMixin,
+    tooltipsMixin,
   ],
 
   components: {
@@ -87,10 +82,6 @@ export default {
     RegistrationMenu,
     TeamMenu,
     JudgingMenu,
-  },
-
-  created () {
-    this.Tooltips = Tooltips
   },
 
   computed: {
@@ -152,12 +143,12 @@ export default {
 
     submissionDisabledTooltipMessage () {
       if (!this.isOnTeam && !this.isConsentSigned)
-        return Tooltips.MUST_SIGN_CONSENT_ON_TEAM
+        return Tooltips.mentor.submissions.MUST_SIGN_CONSENT_ON_TEAM
 
       if (!this.isConsentSigned)
-        return Tooltips.MUST_SIGN_CONSENT
+        return Tooltips.mentor.submissions.MUST_SIGN_CONSENT
 
-      return Tooltips.MUST_BE_ON_TEAM
+      return Tooltips.mentor.submissions.MUST_BE_ON_TEAM
     },
   },
 }
